requirements.txt
tensorflow==2.12.1
pytorch==1.13.0
transformers==5.1.1
gitpython==4.1.0
